refactor: Added Segregated cpp_benchmark from examples
