+++ examples/training/vgg16/README.md
