+++ examples/training/vgg16/export_ckpt.py
