+++ examples/training/vgg16/main.py
