+++ examples/training/vgg16/requirements.txt
