+++ examples/training/vgg16/vgg16.py
