+++ tools/cpp_benchmark/int8/datasets/cifar10.h
