+++ tools/cpp_benchmark/int8/qat/README.md
@@ -12,7 +12,7 @@ This is a short example application that shows how to use Torch-TensorRT to perf
 ## Prerequisites
 
 1. Download CIFAR10 Dataset Binary version ([https://www.cs.toronto.edu/~kriz/cifar-10-binary.tar.gz](https://www.cs.toronto.edu/~kriz/cifar-10-binary.tar.gz))
-2. Train a network on CIFAR10 and perform quantization aware training on it. Refer to `examples/int8/training/vgg16/README.md` for detailed instructions.
+2. Train a network on CIFAR10 and perform quantization aware training on it. Refer to `examples/training/vgg16/README.md` for detailed instructions.
    Export the QAT model to Torchscript.
 3. Install NVIDIA's <a href="https://github.com/NVIDIA/TensorRT/tree/master/tools/pytorch-quantization">pytorch quantization toolkit</a>
 4. TensorRT 8.0.1.6 or above
@@ -20,13 +20,13 @@ This is a short example application that shows how to use Torch-TensorRT to perf
 ## Compilation using bazel
 
 ``` shell
-bazel run //examples/int8/qat --compilation_mode=opt <path-to-module> <path-to-cifar10>
+bazel run //tools/cpp_benchmark/int8/qat --compilation_mode=opt <path-to-module> <path-to-cifar10>
 ```
 
 If you want insight into what is going under the hood or need debug symbols
 
 ``` shell
-bazel run //examples/int8/qat --compilation_mode=dbg <path-to-module> <path-to-cifar10>
+bazel run //tools/cpp_benchmark/int8/qat --compilation_mode=dbg <path-to-module> <path-to-cifar10>
 ```
 
 This will build a binary named `qat` in `bazel-out/k8-<opt|dbg>/bin/cpp/int8/qat/` directory. Optionally you can add this to `$PATH` environment variable to run `qat` from anywhere on your system.
@@ -60,7 +60,7 @@ We import header files `cifar10.h` and `benchmark.h` from `ROOT_DIR`. `ROOT_DIR`
 By default it is set to `../../../`. If your Torch-TensorRT directory structure is different, please set `ROOT_DIR` accordingly.
 
 ```sh
-cd examples/int8/qat
+cd tools/cpp_benchmark/int8/qat
 # This will generate a ptq binary
 make ROOT_DIR=<PATH> CUDA_VERSION=11.1
 ./qat <path-to-module> <path-to-cifar10>
